new Score description file
Add new idea of scores in game. Add step to next level.

+++ labirynth/src/com/flexymind/labirynth/objects/Square.java
@@ -140,12 +140,12 @@ public class Square {
 		textureBuffer.position(0);
 	}
 	
-	public int getWidth(){
-		return (int)(width);
+	public float getWidth(){
+		return width;
 	}
 	
-	public int getHeight(){
-		return (int)(height);
+	public float getHeight(){
+		return height;
 	}
 	
 	public void moveTo(PointF vec){
